requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.6
langchain-core==0.3.15
langchain-ollama==0.2.0
langchain-text-splitters==0.3.1
langsmith==0.1.139
Markdown==3.7
multidict==6.1.0
numpy==1.26.4
ollama==0.3.3
orjson==3.10.10
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
